+++ src/GUI/DropdownList.cpp
@@ -3,7 +3,12 @@
 using namespace qolmod;
 
 /*
-auto test = DropdownList::create({DropdownItem({"questsBtn.png"_spr, "i send you on a quest to find yuri"}), DropdownItem({"heart.png"_spr, "yuri yuri yuri"}), DropdownItem({"keybinds.png"_spr, "use this keyboard to find yuri"}), DropdownItem({"discord.png"_spr, "send me yuri on this platform"})});
+auto test = DropdownList::create({
+    DropdownItem({"questsBtn.png"_spr, "i send you on a quest to find yuri"}),
+    DropdownItem({"heart.png"_spr, "yuri yuri yuri"}),
+    DropdownItem({"keybinds.png"_spr, "use this keyboard to find yuri"}),
+    DropdownItem({"discord.png"_spr, "send me yuri on this platform"})
+    });
 test->setPosition(ccp(150, 150));
 
 this->addChild(test);